requirements.txt
fastapi
uvicorn
pillow
numpy
requests
openai
mediapipe
firebase-admin
opencv-python
protobuf==3.20.0
